fix bug when job runner submit stdout doesn't end with newline
When the stdout returned by the job runner submit method doesn't end with
a newline, the `sys.stdout.write` call in `jobs_submit` do not add it and
this cases submission to fails (at least to be recorded as failed by the
calling process when running a workflow).

Looking at how newlines are handled in the JobRunnerManager.job_kill method,
it looks like this a typo in the jobs_submit method.